Extended the test suite to look at mutation operations as well.  Also, added a shell script to automate testing over major implementations.
At this point it seems that IMA doesn't work with CMUCL.  It should, seems like
a CMUCL bug actually.  Maybe my install is corrupt in some way?

It also doesn't work with ABCL, but that is due to ABCL not
supporting (setf (apply user-function args) ...) forms like the spec says it
should.