Grouping and rasing dimensionalities of arrays is partially working.
We can group, but only using 1D imas as the grouper (the grouped imas have be
whatever)

We can raise-dimensionality, but only on the last index

We can add an extent=1 index anywhere we want.

Also:

Fixed a but in def-generic-map, moved around some code so it flows better in the
document.  Had to comment map-ima as it was failing to compile as the iterate
extensions aren't in place yet.